Injection Plugin not installing correctly
Hi, it’s possible you may have to remove my plugin as it is not installing properly in Alcatraz since changes I made yesterday. It now has a bundle subproject meaning the Plugin project needs to be built for the "iOS simulator". Are there Alcatraz logs I can look at to find out what the problem is?